stackage: move all source files out of extension-package
It is a bit inconvenient to manage the `extension-package` folder when
some of it is source files and some of it is generated files, so this
changeset moves all of the source files into other places. Now,
`extension-package` will be fully generated, and can be treated as a
build output folder.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>